ai-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
pytesseract==0.3.13
pillow==11.1.0
scikit-learn==1.6.1
pandas==2.2.3
numpy==1.26.4
easyocr==1.7.2
opencv-python==4.11.0.86
transformers==4.49.0
torch==2.6.0
prophet==1.1.6
tensorflow==2.19.0
joblib==1.4.2
stanza==1.8.0
spacy==3.7.2
nltk==3.8.1
redis==5.0.1
